+++ babyphone/babyphone/src/includes/ppm_out.h
@@ -8,10 +8,10 @@
 
 #ifndef PPM_OUT_H_
 #define PPM_OUT_H_
-
+#include "stdint.h"
 
 #define PIN_PPM_OUT     PIN_TC0_TIOA0  // PIN TIOA1 which is PIN PA2 resp. "A7" on Due
 void ppm_out_initialize(void);
-void set_ppm_out_channel_value(int idx, int value);
+void set_ppm_out_channel_value(int idx, uint16_t value);
 
 #endif /* PPM_OUT_H_ */
